+++ crates/grafana-plugin-sdk-macros/tests/ui-fail/non_impl_service.rs
@@ -1,7 +1,10 @@
 #![allow(dead_code, unused_variables)]
 
 mod a {
-    #[derive(Clone)]
+    use grafana_plugin_sdk::prelude::*;
+
+    #[derive(Clone, GrafanaPlugin)]
+    #[grafana_plugin(plugin_type = "datasource")]
     struct MyPlugin;
 
     #[grafana_plugin_sdk::main(services(data))]
@@ -11,5 +14,3 @@ mod a {
 }
 
 fn main() {}
-
-